# daemon/openapi.yaml
...
openapi: 3.0.0
info:
  version: 0.1.0
  title: Grid REST API
  description: _An API providing HTTP/JSON interface to Hyperledger Grid._
paths:
...
  "/record/{record_id}":
    get:
      tags:
        - Track and Trace
      summary: Fetch a particular record
      description: Fetches a single record with the given record ID.
      operationId: fetch_record
      parameters:
        - name: record_id
          in: path
          description: ID of the record to return
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: "#/components/schemas/Record"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "404":
          $ref: "#/components/responses/404NotFound"
        "500":
          $ref: "#/components/responses/500ServerError"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
  "/record/{record_id}/property/{property_name}":
      get:
        tags:
          - Track and Trace
        summary: Fetch a particular property
        description: |
          Fetches a single property with the given record ID and property name.
        operationId: fetch_property
        parameters:
          - name: record_id
            in: path
            description: ID of the record to fetch a property from.
            required: true
            schema:
              type: string
          - name: property_name
            in: path
            description: Name of the property to fetch.
            required: true
            schema:
              type: string
        responses:
          "200":
            description: Successful operation
            content:
              application/json:
                schema:
                  properties:
                    data:
                      $ref: "#/components/schemas/Property"
          "400":
            $ref: "#/components/responses/400BadRequest"
          "404":
            $ref: "#/components/responses/404NotFound"
          "500":
            $ref: "#/components/responses/500ServerError"
          "503":
            $ref: "#/components/responses/503ServiceUnavailable"
components:
...
  responses:
    400BadRequest:
      description: Request was malformed
      content:
        "*/*":
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Address or id did not match any resource
      content:
        "*/*":
          schema:
            $ref: "#/components/schemas/Error"
...
    500ServerError:
      description: Something went wrong within the validator
      content:
        "*/*":
          schema:
            $ref: "#/components/schemas/Error"
    503ServiceUnavailable:
      description: API is unable to reach the validator
      content:
        "*/*":
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      properties:
        code:
          type: integer
          example: 34
        title:
          type: string
          example: No Batches Submitted
        message:
          type: string
          example: >
            The protobuf BatchList you submitted was empty and contained no
            Batches. You must submit at least one Batch.
...
    DataTypeEnum:
      description: Data type of a PropertyDefinition
      type: string
      enum:
        - BYTES
        - BOOLEAN
        - NUMBER
        - STRING
        - ENUM
        - STRUCT
        - LOCATION
...
    AssociatedAgent:
      type: object
      properties:
        agent_id:
          type: string
          example: 02fb5b3a093e20e420ecf9c5839215e74c97f49eb51889069eb87bc6f62ceca8dd
        timestamp:
          type: integer
          example: 1557949075
    LatLong:
      type: object
      properties:
        latitude:
          type: integer
          example: 46786299
        longitude:
          type: integer
          example: -92051336
    Record:
      type: object
      properties:
        record_id:
          type: string
          example: 7h15-45537-15-br173
        owner:
          type: string
          example: 02cd3181dbd7d1539f470436ce222c53ab5e514f67809dc0095895e6cdfba97612
        custodian:
          type: string
          example: 02fb5b3a093e20e420ecf9c5839215e74c97f49eb51889069eb87bc6f62ceca8dd
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        proposals:
          type: array
          items:
            $ref: "#/components/schemas/Proposal"
        owner_updates:
          type: array
          items:
            $ref: "#/components/schemas/AssociatedAgent"
        custodian_updates:
          type: array
          items:
            $ref: "#/components/schemas/AssociatedAgent"
        final:
          type: boolean
    ReportedValue:
      type: object
      properties:
        timestamp:
          type: integer
          example: 1557949075
        value:
          oneOf:
            - $ref: '#/components/schemas/StringValue'
            - $ref: '#/components/schemas/BoolValue'
            - $ref: '#/components/schemas/StructValue'
            - $ref: '#/components/schemas/LocationValue'
            - $ref: '#/components/schemas/NumberValue'
            - $ref: '#/components/schemas/EnumValue'
            - $ref: '#/components/schemas/BytesValue'
          example: {Bytes: "AQIDBA=="}
        reporter:
          type: object
          properties:
            metadata:
              type: object
              example: {agent_name: "Smith"}
            public_key:
              type: string
              example: 02cd3181dbd7d1539f470436ce222c53ab5e514f67809dc0095895e6cdfba97612
    Property:
      type: object
      properties:
        name:
          type: string
          example: location
        record_id:
          type: string
          example: 7h15-45537-15-br173
        data_type:
          $ref: "#/components/schemas/DataTypeEnum"
        authorized_reporters:
          type: array
          items:
            type: string
            example:
              - 02cd3181dbd7d1539f470436ce222c53ab5e514f67809dc0095895e6cdfba97612
              - 0364edd42bd9b2dea1315e2da820b569665f96e36c44b267ceeac488cfdc03bf61
        value:
            $ref: "#/components/schemas/ReportedValue"
        updates:
          type: array
          items:
            $ref: "#/components/schemas/ReportedValue"
    ProposalRoleEnum:
      type: string
      enum:
        - OWNER
        - CUSTODIAN
        - REPORTER
    ProposalStatusEnum:
      description: Status of a proposal
      type: string
      enum:
        - OPEN
        - ACCEPTED
        - REJECTED
        - CANCELED
    Proposal:
      type: object
      properties:
        receiving_agent:
          type: string
          example: 02cd3181dbd7d1539f470436ce222c53ab5e514f67809dc0095895e6cdfba97612
        issuing_agent:
          type: string
          example: 02cd3181dbd7d1539f470436ce222c53ab5e514f67809dc0095895e6cdfba97612
        role:
          $ref: "#/components/schemas/ProposalRoleEnum"
        properties:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ProposalStatusEnum"
        timestamp:
          type: integer
          example: 1557949075
    StructPropertyValue:
      type: object
      properties:
        name:
          type: string
          example: location
        type:
          $ref: "#/components/schemas/DataTypeEnum"
        value:
          oneOf:
            - $ref: '#/components/schemas/StringValue'
            - $ref: '#/components/schemas/BoolValue'
            - $ref: '#/components/schemas/StructValue'
            - $ref: '#/components/schemas/LocationValue'
            - $ref: '#/components/schemas/NumberValue'
            - $ref: '#/components/schemas/EnumValue'
            - $ref: '#/components/schemas/BytesValue'
          example: {String: lighbulb}
    StringValue:
      type: object
      properties:
        String:
          type: string
          example: "lightbulb"
    BoolValue:
      type: object
      properties:
        Bool:
          type: boolean
          example: true
    StructValue:
      type: object
      properties:
        Struct:
          type: array
          items:
            $ref: '#/components/schemas/StructPropertyValue'
    LocationValue:
      type: object
      properties:
        Location:
          $ref: '#/components/schemas/LatLong'
    NumberValue:
       type: object
       properties:
         Number:
           type: integer
           format: int64
           example: 2
    EnumValue:
      type: object
      properties:
        Enum:
          type: integer
          format: int32
          example: 32
    BytesValue:
       type: object
       properties:
         Bytes:
          type: string
          format: byte
          example: "AQIDBA=="